data-augmentation/requirements.txt
Pillow
numpy
torch
torchvision
torchaudio
ultralytics
python-dotenv